+++ linuxcnc-hal/src/hal_pin/bidirectional_pin.rs
@@ -3,24 +3,29 @@ use crate::{
     hal_pin::{HalPin, PinDirection},
 };
 
-/// Wrapping struct to specialise a HAL pin to an output
-pub struct OutputPin<P> {
+/// Wrapping struct to specialise a HAL pin to a bidirectional IO
+pub struct BidirectionalPin<P> {
     pin: P,
 }
 
-impl<P> OutputPin<P>
+impl<P> BidirectionalPin<P>
 where
     P: HalPin,
 {
-    /// Register a new output pin with the HAL
+    /// Register a new bidirectional pin with the HAL
     ///
     /// Requires the full pin name including component like `vfd.spindle-speed-out` or
     /// `jog-pendant.is-estopped`. The component ID should be fetched from
     /// [`HalComponentBuilder.id`].
     pub fn new(name: String, component_id: i32) -> Result<Self, PinRegisterError> {
-        let pin = P::register_pin(&name, PinDirection::Out, component_id)?;
+        let pin = P::register_pin(&name, PinDirection::Bidirectional, component_id)?;
 
         Ok(Self { pin })
+    }
+
+    /// Get the pin's value
+    pub fn value(&self) -> Result<&P::Storage, StorageError> {
+        self.pin.storage()
     }
 
     /// Set the pin's value